Make TypeInstrumentations top-level classes
