+++ instrumentation/apache-httpclient/apache-httpclient-4.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/apachehttpclient/v4_0/ApacheHttpClientInstrumentationModule.java
@@ -5,32 +5,12 @@
 
 package io.opentelemetry.javaagent.instrumentation.apachehttpclient.v4_0;
 
-import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.implementsInterface;
-import static io.opentelemetry.javaagent.extension.matcher.ClassLoaderMatcher.hasClassesNamed;
-import static io.opentelemetry.javaagent.instrumentation.apachehttpclient.v4_0.ApacheHttpClientInstrumenters.instrumenter;
-import static io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge.currentContext;
 import static java.util.Collections.singletonList;
-import static net.bytebuddy.matcher.ElementMatchers.isAbstract;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.not;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import com.google.auto.service.AutoService;
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
 import java.util.List;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
-import org.apache.http.HttpHost;
-import org.apache.http.HttpRequest;
-import org.apache.http.client.ResponseHandler;
-import org.apache.http.client.methods.HttpUriRequest;
 
 @AutoService(InstrumentationModule.class)
 public class ApacheHttpClientInstrumentationModule extends InstrumentationModule {
@@ -41,250 +21,6 @@ public class ApacheHttpClientInstrumentationModule extends InstrumentationModule
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new HttpClientInstrumentation());
-  }
-
-  public static class HttpClientInstrumentation implements TypeInstrumentation {
-    @Override
-    public ElementMatcher<ClassLoader> classLoaderOptimization() {
-      return hasClassesNamed("org.apache.http.client.HttpClient");
-    }
-
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      return implementsInterface(named("org.apache.http.client.HttpClient"));
-    }
-
-    @Override
-    public void transform(TypeTransformer transformer) {
-      // There are 8 execute(...) methods.  Depending on the version, they may or may not delegate
-      // to each other. Thus, all methods need to be instrumented.  Because of argument position and
-      // type, some methods can share the same advice class.  The call depth tracking ensures only 1
-      // span is created
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(1))
-              .and(takesArgument(0, named("org.apache.http.client.methods.HttpUriRequest"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$UriRequestAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(2))
-              .and(takesArgument(0, named("org.apache.http.client.methods.HttpUriRequest")))
-              .and(takesArgument(1, named("org.apache.http.protocol.HttpContext"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$UriRequestAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(2))
-              .and(takesArgument(0, named("org.apache.http.client.methods.HttpUriRequest")))
-              .and(takesArgument(1, named("org.apache.http.client.ResponseHandler"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$UriRequestWithHandlerAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(3))
-              .and(takesArgument(0, named("org.apache.http.client.methods.HttpUriRequest")))
-              .and(takesArgument(1, named("org.apache.http.client.ResponseHandler")))
-              .and(takesArgument(2, named("org.apache.http.protocol.HttpContext"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$UriRequestWithHandlerAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(2))
-              .and(takesArgument(0, named("org.apache.http.HttpHost")))
-              .and(takesArgument(1, named("org.apache.http.HttpRequest"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$RequestAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(3))
-              .and(takesArgument(0, named("org.apache.http.HttpHost")))
-              .and(takesArgument(1, named("org.apache.http.HttpRequest")))
-              .and(takesArgument(2, named("org.apache.http.protocol.HttpContext"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$RequestAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(3))
-              .and(takesArgument(0, named("org.apache.http.HttpHost")))
-              .and(takesArgument(1, named("org.apache.http.HttpRequest")))
-              .and(takesArgument(2, named("org.apache.http.client.ResponseHandler"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$RequestWithHandlerAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("execute"))
-              .and(not(isAbstract()))
-              .and(takesArguments(4))
-              .and(takesArgument(0, named("org.apache.http.HttpHost")))
-              .and(takesArgument(1, named("org.apache.http.HttpRequest")))
-              .and(takesArgument(2, named("org.apache.http.client.ResponseHandler")))
-              .and(takesArgument(3, named("org.apache.http.protocol.HttpContext"))),
-          ApacheHttpClientInstrumentationModule.class.getName() + "$RequestWithHandlerAdvice");
-    }
-  }
-
-  public static class UriRequestAdvice {
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void methodEnter(
-        @Advice.Argument(0) HttpUriRequest request,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-      if (!instrumenter().shouldStart(parentContext, request)) {
-        return;
-      }
-
-      context = instrumenter().start(parentContext, request);
-      scope = context.makeCurrent();
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void methodExit(
-        @Advice.Argument(0) HttpUriRequest request,
-        @Advice.Return Object result,
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      ApacheHttpClientHelper.doMethodExit(context, request, result, throwable);
-    }
-  }
-
-  public static class UriRequestWithHandlerAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void methodEnter(
-        @Advice.Argument(0) HttpUriRequest request,
-        @Advice.Argument(value = 1, readOnly = false) ResponseHandler<?> handler,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-      if (!instrumenter().shouldStart(parentContext, request)) {
-        return;
-      }
-
-      context = instrumenter().start(parentContext, request);
-      scope = context.makeCurrent();
-
-      // Wrap the handler so we capture the status code
-      if (handler != null) {
-        handler =
-            new WrappingStatusSettingResponseHandler<>(context, parentContext, request, handler);
-      }
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void methodExit(
-        @Advice.Argument(0) HttpUriRequest request,
-        @Advice.Return Object result,
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      ApacheHttpClientHelper.doMethodExit(context, request, result, throwable);
-    }
-  }
-
-  public static class RequestAdvice {
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void methodEnter(
-        @Advice.Argument(0) HttpHost host,
-        @Advice.Argument(1) HttpRequest request,
-        @Advice.Local("fullRequest") HttpUriRequest fullRequest,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-
-      fullRequest = new HostAndRequestAsHttpUriRequest(host, request);
-
-      if (!instrumenter().shouldStart(parentContext, fullRequest)) {
-        return;
-      }
-
-      context = instrumenter().start(parentContext, fullRequest);
-      scope = context.makeCurrent();
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void methodExit(
-        @Advice.Return Object result,
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("fullRequest") HttpUriRequest fullRequest,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      ApacheHttpClientHelper.doMethodExit(context, fullRequest, result, throwable);
-    }
-  }
-
-  public static class RequestWithHandlerAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void methodEnter(
-        @Advice.Argument(0) HttpHost host,
-        @Advice.Argument(1) HttpRequest request,
-        @Advice.Argument(value = 2, readOnly = false) ResponseHandler<?> handler,
-        @Advice.Local("fullRequest") HttpUriRequest fullRequest,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-      if (!instrumenter().shouldStart(parentContext, fullRequest)) {
-        return;
-      }
-
-      context = instrumenter().start(parentContext, fullRequest);
-      scope = context.makeCurrent();
-
-      // Wrap the handler so we capture the status code
-      if (handler != null) {
-        handler =
-            new WrappingStatusSettingResponseHandler<>(
-                context, parentContext, fullRequest, handler);
-      }
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void methodExit(
-        @Advice.Return Object result,
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("fullRequest") HttpUriRequest fullRequest,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      ApacheHttpClientHelper.doMethodExit(context, fullRequest, result, throwable);
-    }
+    return singletonList(new ApacheHttpClientInstrumentation());
   }
 }
