+++ instrumentation/dropwizard-views-0.7/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/dropwizardviews/DropwizardRendererInstrumentation.java
@@ -8,56 +8,40 @@ package io.opentelemetry.javaagent.instrumentation.dropwizardviews;
 import static io.opentelemetry.javaagent.extension.matcher.AgentElementMatchers.implementsInterface;
 import static io.opentelemetry.javaagent.extension.matcher.ClassLoaderMatcher.hasClassesNamed;
 import static io.opentelemetry.javaagent.instrumentation.dropwizardviews.DropwizardTracer.tracer;
-import static java.util.Collections.singletonList;
 import static net.bytebuddy.matcher.ElementMatchers.isMethod;
 import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 
-import com.google.auto.service.AutoService;
 import io.dropwizard.views.View;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
-import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
 import io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge;
-import java.util.List;
 import net.bytebuddy.asm.Advice;
 import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
-@AutoService(InstrumentationModule.class)
-public class DropwizardViewsInstrumentationModule extends InstrumentationModule {
-  public DropwizardViewsInstrumentationModule() {
-    super("dropwizard-views");
+public class DropwizardRendererInstrumentation implements TypeInstrumentation {
+  @Override
+  public ElementMatcher<ClassLoader> classLoaderOptimization() {
+    return hasClassesNamed("io.dropwizard.views.ViewRenderer");
   }
 
   @Override
-  public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new ViewRendererInstrumentation());
+  public ElementMatcher<TypeDescription> typeMatcher() {
+    return implementsInterface(named("io.dropwizard.views.ViewRenderer"));
   }
 
-  public static class ViewRendererInstrumentation implements TypeInstrumentation {
-    @Override
-    public ElementMatcher<ClassLoader> classLoaderOptimization() {
-      return hasClassesNamed("io.dropwizard.views.ViewRenderer");
-    }
-
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      return implementsInterface(named("io.dropwizard.views.ViewRenderer"));
-    }
-
-    @Override
-    public void transform(TypeTransformer transformer) {
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("render"))
-              .and(takesArgument(0, named("io.dropwizard.views.View")))
-              .and(isPublic()),
-          DropwizardViewsInstrumentationModule.class.getName() + "$RenderAdvice");
-    }
+  @Override
+  public void transform(TypeTransformer transformer) {
+    transformer.applyAdviceToMethod(
+        isMethod()
+            .and(named("render"))
+            .and(takesArgument(0, named("io.dropwizard.views.View")))
+            .and(isPublic()),
+        this.getClass().getName() + "$RenderAdvice");
   }
 
   public static class RenderAdvice {
