+++ instrumentation/google-http-client-1.19/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/googlehttpclient/GoogleHttpClientInstrumentationModule.java
@@ -5,29 +5,12 @@
 
 package io.opentelemetry.javaagent.instrumentation.googlehttpclient;
 
-import static io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge.currentContext;
-import static io.opentelemetry.javaagent.instrumentation.googlehttpclient.GoogleHttpClientTracer.tracer;
 import static java.util.Collections.singletonList;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.isPublic;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
-import com.google.api.client.http.HttpRequest;
-import com.google.api.client.http.HttpResponse;
 import com.google.auto.service.AutoService;
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import io.opentelemetry.javaagent.instrumentation.api.InstrumentationContext;
 import java.util.List;
-import java.util.concurrent.Executor;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
-import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(InstrumentationModule.class)
 public class GoogleHttpClientInstrumentationModule extends InstrumentationModule {
@@ -37,103 +20,6 @@ public class GoogleHttpClientInstrumentationModule extends InstrumentationModule
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new HttpRequestInstrumentation());
-  }
-
-  public static class HttpRequestInstrumentation implements TypeInstrumentation {
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      // HttpRequest is a final class.  Only need to instrument it exactly
-      // Note: the rest of com.google.api is ignored in AdditionalLibraryIgnoresMatcher to speed
-      // things up
-      return named("com.google.api.client.http.HttpRequest");
-    }
-
-    @Override
-    public void transform(TypeTransformer transformer) {
-      transformer.applyAdviceToMethod(
-          isMethod().and(isPublic()).and(named("execute")).and(takesArguments(0)),
-          GoogleHttpClientInstrumentationModule.class.getName() + "$GoogleHttpClientAdvice");
-
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(isPublic())
-              .and(named("executeAsync"))
-              .and(takesArguments(1))
-              .and(takesArgument(0, Executor.class)),
-          GoogleHttpClientInstrumentationModule.class.getName() + "$GoogleHttpClientAsyncAdvice");
-    }
-  }
-
-  public static class GoogleHttpClientAdvice {
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void methodEnter(
-        @Advice.This HttpRequest request,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-
-      context = InstrumentationContext.get(HttpRequest.class, Context.class).get(request);
-      if (context != null) {
-        // span was created by GoogleHttpClientAsyncAdvice instrumentation below
-        // (executeAsync ends up calling execute from a separate thread)
-        // so make it current and end it in method exit
-        scope = context.makeCurrent();
-        return;
-      }
-      Context parentContext = currentContext();
-      if (!tracer().shouldStartSpan(parentContext)) {
-        return;
-      }
-      context = tracer().startSpan(parentContext, request);
-      scope = context.makeCurrent();
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void methodExit(
-        @Advice.Return HttpResponse response,
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      tracer().endMaybeExceptionally(context, response, throwable);
-    }
-  }
-
-  public static class GoogleHttpClientAsyncAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void methodEnter(
-        @Advice.This HttpRequest request,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-      if (!tracer().shouldStartSpan(parentContext)) {
-        return;
-      }
-
-      context = tracer().startSpan(parentContext, request);
-      scope = context.makeCurrent();
-
-      InstrumentationContext.get(HttpRequest.class, Context.class).put(request, context);
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void methodExit(
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      if (throwable != null) {
-        tracer().endExceptionally(context, throwable);
-      }
-    }
+    return singletonList(new GoogleHttpRequestInstrumentation());
   }
 }
