+++ instrumentation/gwt-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/gwt/GwtInstrumentationModule.java
@@ -6,22 +6,12 @@
 package io.opentelemetry.javaagent.instrumentation.gwt;
 
 import static io.opentelemetry.javaagent.extension.matcher.ClassLoaderMatcher.hasClassesNamed;
-import static io.opentelemetry.javaagent.instrumentation.gwt.GwtTracer.tracer;
 import static java.util.Collections.singletonList;
-import static net.bytebuddy.matcher.ElementMatchers.named;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import com.google.auto.service.AutoService;
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
-import java.lang.reflect.Method;
 import java.util.List;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
 
 @AutoService(InstrumentationModule.class)
@@ -39,70 +29,6 @@ public class GwtInstrumentationModule extends InstrumentationModule {
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new RpcInstrumentation());
-  }
-
-  public static class RpcInstrumentation implements TypeInstrumentation {
-
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      return named("com.google.gwt.user.server.rpc.RPC");
-    }
-
-    @Override
-    public void transform(TypeTransformer transformer) {
-      transformer.applyAdviceToMethod(
-          named("invokeAndEncodeResponse")
-              .and(takesArguments(5))
-              .and(takesArgument(0, Object.class))
-              .and(takesArgument(1, Method.class))
-              .and(takesArgument(2, Object[].class))
-              .and(takesArgument(3, named("com.google.gwt.user.server.rpc.SerializationPolicy")))
-              .and(takesArgument(4, int.class)),
-          RpcInstrumentation.class.getName() + "$RpcInvokeAdvice");
-
-      // encodeResponseForFailure is called by invokeAndEncodeResponse in case of failure
-      transformer.applyAdviceToMethod(
-          named("encodeResponseForFailure")
-              .and(takesArguments(4))
-              .and(takesArgument(0, Method.class))
-              .and(takesArgument(1, Throwable.class))
-              .and(takesArgument(2, named("com.google.gwt.user.server.rpc.SerializationPolicy")))
-              .and(takesArgument(3, int.class)),
-          RpcInstrumentation.class.getName() + "$RpcFailureAdvice");
-    }
-
-    public static class RpcInvokeAdvice {
-      @Advice.OnMethodEnter(suppress = Throwable.class)
-      public static void onEnter(
-          @Advice.Argument(0) Object target,
-          @Advice.Argument(1) Method method,
-          @Advice.Local("otelContext") Context context,
-          @Advice.Local("otelScope") Scope scope) {
-
-        context = tracer().startRpcSpan(target, method);
-        scope = context.makeCurrent();
-      }
-
-      @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-      public static void onExit(
-          @Advice.Thrown Throwable throwable,
-          @Advice.Local("otelContext") Context context,
-          @Advice.Local("otelScope") Scope scope) {
-        scope.close();
-
-        tracer().endSpan(context, throwable);
-      }
-    }
-
-    public static class RpcFailureAdvice {
-      @Advice.OnMethodEnter(suppress = Throwable.class)
-      public static void onEnter(@Advice.Argument(1) Throwable throwable) {
-        if (throwable == null) {
-          return;
-        }
-        tracer().rpcFailure(throwable);
-      }
-    }
+    return singletonList(new GwtRpcInstrumentation());
   }
 }
