+++ instrumentation/jedis/jedis-1.4/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/jedis/v1_4/JedisInstrumentationModule.java
@@ -6,29 +6,14 @@
 package io.opentelemetry.javaagent.instrumentation.jedis.v1_4;
 
 import static io.opentelemetry.javaagent.extension.matcher.ClassLoaderMatcher.hasClassesNamed;
-import static io.opentelemetry.javaagent.instrumentation.api.Java8BytecodeBridge.currentContext;
-import static io.opentelemetry.javaagent.instrumentation.jedis.v1_4.JedisInstrumenters.instrumenter;
-import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
-import static net.bytebuddy.matcher.ElementMatchers.is;
-import static net.bytebuddy.matcher.ElementMatchers.isMethod;
-import static net.bytebuddy.matcher.ElementMatchers.named;
 import static net.bytebuddy.matcher.ElementMatchers.not;
-import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
 import com.google.auto.service.AutoService;
-import io.opentelemetry.context.Context;
-import io.opentelemetry.context.Scope;
 import io.opentelemetry.javaagent.extension.instrumentation.InstrumentationModule;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
-import io.opentelemetry.javaagent.extension.instrumentation.TypeTransformer;
 import java.util.List;
-import net.bytebuddy.asm.Advice;
-import net.bytebuddy.description.type.TypeDescription;
 import net.bytebuddy.matcher.ElementMatcher;
-import redis.clients.jedis.Connection;
-import redis.clients.jedis.Protocol.Command;
 
 @AutoService(InstrumentationModule.class)
 public class JedisInstrumentationModule extends InstrumentationModule {
@@ -45,101 +30,6 @@ public class JedisInstrumentationModule extends InstrumentationModule {
 
   @Override
   public List<TypeInstrumentation> typeInstrumentations() {
-    return singletonList(new ConnectionInstrumentation());
-  }
-
-  public static class ConnectionInstrumentation implements TypeInstrumentation {
-    @Override
-    public ElementMatcher<TypeDescription> typeMatcher() {
-      return named("redis.clients.jedis.Connection");
-    }
-
-    @Override
-    public void transform(TypeTransformer transformer) {
-      // FIXME: This instrumentation only incorporates sending the command, not processing the
-      // result.
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("sendCommand"))
-              .and(takesArguments(1))
-              .and(takesArgument(0, named("redis.clients.jedis.Protocol$Command"))),
-          JedisInstrumentationModule.class.getName() + "$JedisNoArgsAdvice");
-      transformer.applyAdviceToMethod(
-          isMethod()
-              .and(named("sendCommand"))
-              .and(takesArguments(2))
-              .and(takesArgument(0, named("redis.clients.jedis.Protocol$Command")))
-              .and(takesArgument(1, is(byte[][].class))),
-          JedisInstrumentationModule.class.getName() + "$JedisArgsAdvice");
-    }
-  }
-
-  public static class JedisNoArgsAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void onEnter(
-        @Advice.This Connection connection,
-        @Advice.Argument(0) Command command,
-        @Advice.Local("otelJedisRequest") JedisRequest request,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-      request = JedisRequest.create(connection, command);
-      if (!instrumenter().shouldStart(parentContext, request)) {
-        return;
-      }
-
-      context = instrumenter().start(parentContext, request);
-      scope = context.makeCurrent();
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void stopSpan(
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelJedisRequest") JedisRequest request,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      instrumenter().end(context, request, null, throwable);
-    }
-  }
-
-  public static class JedisArgsAdvice {
-
-    @Advice.OnMethodEnter(suppress = Throwable.class)
-    public static void onEnter(
-        @Advice.This Connection connection,
-        @Advice.Argument(0) Command command,
-        @Advice.Argument(1) byte[][] args,
-        @Advice.Local("otelJedisRequest") JedisRequest request,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      Context parentContext = currentContext();
-      request = JedisRequest.create(connection, command, asList(args));
-      if (!instrumenter().shouldStart(parentContext, request)) {
-        return;
-      }
-
-      context = instrumenter().start(parentContext, request);
-      scope = context.makeCurrent();
-    }
-
-    @Advice.OnMethodExit(onThrowable = Throwable.class, suppress = Throwable.class)
-    public static void stopSpan(
-        @Advice.Thrown Throwable throwable,
-        @Advice.Local("otelJedisRequest") JedisRequest request,
-        @Advice.Local("otelContext") Context context,
-        @Advice.Local("otelScope") Scope scope) {
-      if (scope == null) {
-        return;
-      }
-
-      scope.close();
-      instrumenter().end(context, request, null, throwable);
-    }
+    return singletonList(new JedisConnectionInstrumentation());
   }
 }
